+++ crates/fuel-core/src/query/balance.rs
@@ -1,18 +1,19 @@
 use crate::{
-    database::{
-        resource::{
-            AssetQuery,
-            AssetSpendTarget,
-            AssetsQuery,
-        },
-        Database,
-    },
+    database::Database,
     state::IterDirection,
 };
+use asset_query::{
+    AssetQuery,
+    AssetSpendTarget,
+    AssetsQuery,
+};
 use fuel_core_storage::Result as StorageResult;
-use fuel_core_types::fuel_tx::{
-    Address,
-    AssetId,
+use fuel_core_types::{
+    fuel_tx::{
+        Address,
+        AssetId,
+    },
+    services::graphql_api::Balance,
 };
 use itertools::Itertools;
 use std::{
@@ -20,40 +21,14 @@ use std::{
     collections::HashMap,
 };
 
-pub trait BalanceQueryData {
-    type Item<'a>
-    where
-        Self: 'a;
-    type Iter<'a>: Iterator<Item = Self::Item<'a>>
-    where
-        Self: 'a;
-
-    fn balance(
-        &self,
-        owner: Address,
-        asset_id: AssetId,
-    ) -> StorageResult<(Address, u64, AssetId)>;
-
-    fn balances(
-        &self,
-        filter: Address, // owner
-        direction: IterDirection,
-    ) -> StorageResult<Self::Iter<'_>>;
-}
+pub mod asset_query;
 
 pub struct BalanceQueryContext<'a>(pub &'a Database);
 
-impl BalanceQueryData for BalanceQueryContext<'_> {
-    type Item<'a> = StorageResult<(Address, u64, AssetId)> where Self: 'a;
-    type Iter<'a> = Box< dyn Iterator<Item = StorageResult<(Address, u64, AssetId)>> + 'a> where Self: 'a;
-
-    fn balance(
-        &self,
-        owner: Address,
-        asset_id: AssetId,
-    ) -> StorageResult<(Address, u64, AssetId)> {
+impl BalanceQueryContext<'_> {
+    pub fn balance(&self, owner: Address, asset_id: AssetId) -> StorageResult<Balance> {
         let db = self.0;
-        let balance = AssetQuery::new(
+        let amount = AssetQuery::new(
             &owner,
             &AssetSpendTarget::new(asset_id, u64::MAX, u64::MAX),
             None,
@@ -70,33 +45,46 @@ impl BalanceQueryData for BalanceQueryContext<'_> {
             Ok(balance)
         })?;
 
-        Ok((owner, balance, asset_id))
+        Ok(Balance {
+            owner,
+            amount,
+            asset_id,
+        })
     }
 
-    fn balances(
+    pub fn balances(
         &self,
-        filter: Address, // owner
+        owner: Address,
         direction: IterDirection,
-    ) -> StorageResult<Self::Iter<'_>> {
+    ) -> impl Iterator<Item = StorageResult<Balance>> + '_ {
         let db = self.0;
 
-        let owner = filter;
-
         let mut amounts_per_asset = HashMap::new();
+        let mut errors = vec![];
 
         for resource in AssetsQuery::new(&owner, None, None, db).unspent_resources() {
-            let resource = resource?;
-            *amounts_per_asset.entry(*resource.asset_id()).or_default() +=
-                resource.amount();
+            match resource {
+                Ok(resource) => {
+                    *amounts_per_asset.entry(*resource.asset_id()).or_default() +=
+                        resource.amount();
+                }
+                Err(err) => {
+                    errors.push(err);
+                }
+            }
         }
 
         let mut balances = amounts_per_asset
             .into_iter()
-            .map(|(asset_id, amount)| (owner, amount, asset_id))
+            .map(|(asset_id, amount)| Balance {
+                owner,
+                amount,
+                asset_id,
+            })
             .collect_vec();
 
         balances.sort_by(|l, r| {
-            if l.2 < r.2 {
+            if l.amount < r.amount {
                 Ordering::Less
             } else {
                 Ordering::Greater
@@ -107,8 +95,9 @@ impl BalanceQueryData for BalanceQueryContext<'_> {
             balances.reverse();
         }
 
-        let balances = balances.into_iter();
-
-        Ok(Box::new(balances))
+        balances
+            .into_iter()
+            .map(Ok)
+            .chain(errors.into_iter().map(Err))
     }
 }
