+++ crates/fuel-core/src/schema/dap.rs
@@ -1,6 +1,6 @@
 use crate::{
     database::{
-        transactional::DatabaseTransaction,
+        transaction::DatabaseTransaction,
         vm_database::VmDatabase,
     },
     fuel_core_graphql_api::service::Database,
