+++ crates/types/src/services.rs
@@ -1,6 +1,7 @@
 //! Types for specific services
 
 pub mod executor;
+pub mod graphql_api;
 pub mod p2p;
 pub mod txpool;
