+++ crates/types/src/services/graphql_api.rs
@@ -1,0 +1,16 @@
+//! Types related to GraphQL API service.
+
+use crate::fuel_types::{
+    Address,
+    AssetId,
+};
+
+/// The cumulative balance(`amount`) of the `owner` of `asset_id`.
+pub struct Balance {
+    /// Owner of the asset.
+    pub owner: Address,
+    /// The cumulative amount of the asset.
+    pub amount: u64,
+    /// The identifier of the asset.
+    pub asset_id: AssetId,
+}
